Fix linting and consolidate unit tests
